refactor(railgun): introduce typed keys and refactor address/key derivation
- Make MasterPublicKey, ViewingPublicKey, NullifyingKey public (with Debug) in crypto::keys and expose to_hex/as_bytes
- Move compute_public_spending_key and compute_public_viewing_key into account/mod.rs (now take path instead of index)
- Update RailgunAddress to use MasterPublicKey + ViewingPublicKey (remove raw U256/[u8;32])
- Refactor RailgunKeys to store typed keys and update derive() to use new SpendingPublicKey + path helpers
- Simplify address.rs: remove heavy ark/curve deps, delete old compute functions, delegate via account/crypto modules
- Update encode_address, decode, and tests to use typed keys and consistent field names (viewing_public_key)
