+++ crates/zeus-railgun/src/account/keys.rs
@@ -1,10 +1,10 @@
-use alloy_primitives::U256;
 use secure_types::SecureArray;
 
-use super::address::{compute_public_spending_key, compute_public_viewing_key};
-use crate::crypto::{
-   keys::{MasterPublicKey, NullifyingKey, SpendingKey, ViewingKey, ViewingPublicKey},
-   poseidon_hash,
+use crate::{
+   account::compute_public_spending_key,
+   crypto::keys::{
+      MasterPublicKey, NullifyingKey, SpendingKey, SpendingPublicKey, ViewingKey, ViewingPublicKey,
+   },
 };
 
 /// BIP-32 spending path
@@ -22,21 +22,25 @@ pub fn viewing_key_path(index: u32) -> String {
 pub struct RailgunKeys {
    /// Spending Private key for signing transactions (BabyJubJub curve).
    pub spending_private_key: SpendingKey,
+
    /// Spending public key point (x, y) on BabyJubJub.
-   pub spending_public: (U256, U256),
+   pub spending_public_key: SpendingPublicKey,
+
    /// Private key for viewing transactions and ECDH.
-   pub viewing_private: ViewingKey,
+   pub viewing_private_key: ViewingKey,
+
    /// Viewing public key (compressed Ed25519 point, 32 bytes) - used in 0zk address.
-   pub viewing_public: [u8; 32],
+   pub viewing_public_key: ViewingPublicKey,
+
    /// Nullifying key = Poseidon(viewing private).
-   pub nullifying_key: U256,
+   pub nullifying_key: NullifyingKey,
+
    /// Master public key = Poseidon(spend_x, spend_y, nullifying_key).
-   pub master_public_key: U256,
+   pub master_public_key: MasterPublicKey,
 }
 
 impl RailgunKeys {
    /// Generate the complete set of Railgun keys (private + public) for a given seed and index.
-   /// This is the recommended entry point for the engine.
    pub fn new(seed: &SecureArray<u8, 64>, index: u32) -> Result<RailgunKeys, anyhow::Error> {
       let spending_path = spending_key_path(index);
       let viewing_path = viewing_key_path(index);
@@ -45,16 +49,21 @@ impl RailgunKeys {
          seed.unlock(|seed_bytes| SpendingKey::derive(seed_bytes, &spending_path))?;
       let viewing_priv = seed.unlock(|seed_bytes| ViewingKey::derive(seed_bytes, &viewing_path))?;
 
-      let (spend_x, spend_y) = compute_public_spending_key(&seed, index)?;
-      let (viewing_public, nullifying_key) = compute_public_viewing_key(&seed, index)?;
+      let (spend_x, spend_y) = compute_public_spending_key(&seed, &spending_path)?;
+      let viewing_public_key = viewing_priv.public_key();
+      let nullifying_key = viewing_priv.nullifying_key();
 
-      let master_public_key = poseidon_hash(&vec![spend_x, spend_y, nullifying_key])?;
+      let x = spend_x.to_be_bytes();
+      let y = spend_y.to_be_bytes();
+      let spending_public_key = SpendingPublicKey::new(x, y);
+
+      let master_public_key = MasterPublicKey::new(spending_public_key, nullifying_key);
 
       Ok(RailgunKeys {
          spending_private_key: spending_priv,
-         spending_public: (spend_x, spend_y),
-         viewing_private: viewing_priv,
-         viewing_public,
+         spending_public_key,
+         viewing_private_key: viewing_priv,
+         viewing_public_key: viewing_public_key,
          nullifying_key,
          master_public_key,
       })
